refactor: Prepare the incoming new DACC service
